ConvexGeometry: Moved to examples
