--- docs/api/math/convexhull/HalfEdge.html
@@ -1,79 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-	<head>
-		<meta charset="utf-8" />
-		<base href="../../../" />
-		<script src="list.js"></script>
-		<script src="page.js"></script>
-		<link type="text/css" rel="stylesheet" href="page.css" />
-	</head>
-	<body>
-		<h1>[name]</h1>
-
-		<div class="desc">
-			The basis for a half-edge data structure, also known as doubly connected edge list (DCEL).<br />
-		</div>
-
-
-		<h2>Constructor</h2>
-
-
-		<h3>[name]( [page:Vertex vertex], [page:Face face] )</h3>
-		[page:Vertex vertex] - [page:Vertex] A reference to its destination vertex.<br /><br />
-		[page:Face face] - [page:Face] A reference to its face.<br />
-
-		</div>
-
-		<h2>Properties</h2>
-
-    <h3>[property:Vertex vertex]</h3>
-    <div>
-      Reference to the destination vertex. The origin vertex can be obtained by querying the destination of its twin, or of the previous half-edge. Default is undefined.
-    </div>
-
-		<h3>[property:HalfEdge prev]</h3>
-		<div>
-			Reference to the previous half-edge of the same face. Default is null.
-		</div>
-
-		<h3>[property:HalfEdge next]</h3>
-		<div>
-			Reference to the next half-edge of the same face. Default is null.
-		</div>
-
-    <h3>[property:HalfEdge twin]</h3>
-    <div>
-      Reference to the twin half-edge to reach the opposite face. Default is null.
-    </div>
-
-		<h3>[property:Face face]</h3>
-		<div>
-			 Each half-edge bounds a single face and thus has a reference to that face. Default is undefined.
-		</div>
-
-    <h2>Methods</h2>
-
-    <h3>[method:Vertex head]()</h3>
-		<div>Returns the destintation vertex.</div>
-
-    <h3>[method:Vertex tail]()</h3>
-    <div>Returns the origin vertex.</div>
-
-    <h3>[method:Float length]()</h3>
-    <div>Returns the [link:https://en.wikipedia.org/wiki/Euclidean_distance Euclidean length]
-		(straight-line length) of the edge.</div>
-
-    <h3>[method:Float lengthSquared]()</h3>
-    <div>Returns the square of the [link:https://en.wikipedia.org/wiki/Euclidean_distance Euclidean length]
-    (straight-line length) of the edge.</div>
-
-    <h3>[method:HalfEdge setTwin]( [page:HalfEdge edge] )</h3>
-    [page:HalfEdge edge] - Any half-edge.<br /><br />
-
-    <div>Sets the twin edge of this half-edge. It also ensures that the twin reference of the given half-edge is correctly set.</div>
-
-		<h2>Source</h2>
-
-		[link:https://github.com/mrdoob/three.js/blob/master/src/[path].js src/[path].js]
-	</body>
-</html>
