+++ docs/examples/geometries/ConvexBufferGeometry.html
@@ -5,17 +5,16 @@
 		<base href="../../" />
 		<script src="list.js"></script>
 		<script src="page.js"></script>
+		<script src="page.js"></script>
 		<link type="text/css" rel="stylesheet" href="page.css" />
 	</head>
 	<body>
-		[page:Geometry] &rarr;
+		[page:BufferGeometry] &rarr;
 
 		<h1>[name]</h1>
 
 		<div class="desc">[name] can be used to generate a convex hull for a given array of 3D points.
 			The average time complexity for this task is considered to be O(nlog(n)).</div>
-
-		<iframe id="scene" src="scenes/geometry-browser.html#ConvexGeometry"></iframe>
 
 		<script>
 
@@ -35,7 +34,7 @@
 
 		<h2>Example</h2>
 
-		<code>var geometry = new THREE.ConvexGeometry( points );
+		<code>var geometry = new THREE.ConvexBufferGeometry( points );
 		var material = new THREE.MeshBasicMaterial( {color: 0x00ff00} );
 		var mesh = new THREE.Mesh( geometry, material );
 		scene.add( mesh );
@@ -51,6 +50,6 @@
 
 		<h2>Source</h2>
 
-		[link:https://github.com/mrdoob/three.js/blob/master/src/[path].js src/[path].js]
+		[link:https://github.com/mrdoob/three.js/blob/master/examples/js/geometries/ConvexGeometry.js examples/js/geometries/ConvexGeometry.js]
 	</body>
 </html>
