+++ docs/examples/geometries/ConvexGeometry.html
@@ -8,14 +8,12 @@
 		<link type="text/css" rel="stylesheet" href="page.css" />
 	</head>
 	<body>
-		[page:BufferGeometry] &rarr;
+		[page:Geometry] &rarr;
 
 		<h1>[name]</h1>
 
 		<div class="desc">[name] can be used to generate a convex hull for a given array of 3D points.
 			The average time complexity for this task is considered to be O(nlog(n)).</div>
-
-		<iframe id="scene" src="scenes/geometry-browser.html#ConvexGeometry"></iframe>
 
 		<script>
 
@@ -35,7 +33,7 @@
 
 		<h2>Example</h2>
 
-		<code>var geometry = new THREE.ConvexBufferGeometry( points );
+		<code>var geometry = new THREE.ConvexGeometry( points );
 		var material = new THREE.MeshBasicMaterial( {color: 0x00ff00} );
 		var mesh = new THREE.Mesh( geometry, material );
 		scene.add( mesh );
@@ -51,6 +49,6 @@
 
 		<h2>Source</h2>
 
-		[link:https://github.com/mrdoob/three.js/blob/master/src/[path].js src/[path].js]
+		[link:https://github.com/mrdoob/three.js/blob/master/examples/js/geometries/ConvexGeometry.js examples/js/geometries/ConvexGeometry.js]
 	</body>
 </html>
