+++ docs/examples/quickhull/Face.html
@@ -1,0 +1,89 @@
+<!DOCTYPE html>
+<html lang="en">
+	<head>
+		<meta charset="utf-8" />
+		<base href="../../" />
+		<script src="list.js"></script>
+		<script src="page.js"></script>
+		<link type="text/css" rel="stylesheet" href="page.css" />
+	</head>
+	<body>
+		<h1>[name]</h1>
+
+		<div class="desc">
+			Represents a section bounded by a specific amount of half-edges. The current implmentation assumes that a face always consist of three edges.
+		</div>
+
+
+		<h2>Constructor</h2>
+
+
+		<h3>[name]()</h3>
+
+		</div>
+
+		<h2>Properties</h2>
+
+		<h3>[property:Vector3 normal]</h3>
+		<div>
+			The normal vector of the face. Default is a [page:Vector3] at (0, 0, 0).
+		</div>
+
+		<h3>[property:Vector3 midpoint]</h3>
+		<div>
+			The midpoint or centroid of the face. Default is a [page:Vector3] at (0, 0, 0).
+		</div>
+
+		<h3>[property:Float area]</h3>
+		<div>
+			The area of the face. Default is 0.
+		</div>
+
+		<h3>[property:Float constant]</h3>
+		<div>
+			Signed distance from face to the origin. Default is 0.
+		</div>
+
+		<h3>[property:Vertex outside]</h3>
+		<div>
+			Reference to a vertex in a vertex list this face can see. Default is null.
+		</div>
+
+		<h3>[property:Integer mark]</h3>
+		<div>
+			Marks if a face is visible or deleted. Default is 'Visible'.
+		</div>
+
+		<h3>[property:HalfEdge edge]</h3>
+		<div>
+			Reference to the base edge of a face. To retrieve all edges, you can use the 'next' reference of the current edge. Default is null.
+		</div>
+
+		<h2>Methods</h2>
+
+		<h3>[method:Face create]( [page:Vertex a], [page:Vertex b], [page:Vertex c] )</h3>
+		[page:Vertex a] - First vertex of the face.<br /><br />
+		[page:Vertex b] - Second vertex of the face.<br /><br />
+		[page:Vertex c] - Third vertex of the face.<br /><br />
+
+		<div>Creates a face.</div>
+
+		<h3>[method:HalfEdge getEdge]( [page:Integer i] )</h3>
+		[page:Integer i] - The index of the edge.<br /><br />
+
+		<div>Returns an edge by the given index.</div>
+
+		<h3>[method:Face compute] ()</h3>
+
+		<div>Computes all properties of the face.</div>
+
+		<h3>[method:Float distanceToPoint]( [page:Vector3 point] )</h3>
+		[page:Vector3 point] - Any point in 3D space.<br /><br />
+
+		<div>Returns the signed distance from a given point to the plane representation of this face.</div>
+
+		<h2>Source</h2>
+
+		[link:https://github.com/mrdoob/three.js/blob/master/examples/js/QuickHull.js examples/js/QuickHull.js]
+	</body>
+</html>
