+++ docs/examples/quickhull/QuickHull.html
@@ -2,7 +2,7 @@
 <html lang="en">
 	<head>
 		<meta charset="utf-8" />
-		<base href="../../../" />
+		<base href="../../" />
 		<script src="list.js"></script>
 		<script src="page.js"></script>
 		<link type="text/css" rel="stylesheet" href="page.css" />
@@ -24,10 +24,10 @@
 
 		<h2>Properties</h2>
 
-    <h3>[property:Float tolerance]</h3>
-    <div>
-      The epsilon value that is used for internal comparative operations. The calculation of this value depends on the size of the geometry. Default is -1.
-    </div>
+		<h3>[property:Float tolerance]</h3>
+		<div>
+			The epsilon value that is used for internal comparative operations. The calculation of this value depends on the size of the geometry. Default is -1.
+		</div>
 
 		<h3>[property:Array faces]</h3>
 		<div>
@@ -54,45 +54,45 @@
 			The internal representation of the given geometry data (an array of [page:Vertex vertices]).
 		</div>
 
-    <h2>Methods</h2>
+		<h2>Methods</h2>
 
-    <h3>[method:QuickHull3 setFromPoints]( [page:Array points] )</h3>
-    [page:Array points] - Array of [page:Vector3 Vector3s] that the resulting convex hull will contain.<br /><br />
+		<h3>[method:QuickHull3 setFromPoints]( [page:Array points] )</h3>
+		[page:Array points] - Array of [page:Vector3 Vector3s] that the resulting convex hull will contain.<br /><br />
 
-    <div>Computes to convex hull for the given array of points.</div>
+		<div>Computes to convex hull for the given array of points.</div>
 
-    <h3>[method:QuickHull3 setFromObject]( [page:Object3D object] )</h3>
-    [page:Object3D object] - [page:Object3D] to compute the convex hull of.<br /><br />
+		<h3>[method:QuickHull3 setFromObject]( [page:Object3D object] )</h3>
+		[page:Object3D object] - [page:Object3D] to compute the convex hull of.<br /><br />
 
-    <div>Computes the convex hull of an [page:Object3D] (including its children),
+		<div>Computes the convex hull of an [page:Object3D] (including its children),
 		accounting for the world transforms of both the object and its childrens.</div>
 
-    <h3>[method:QuickHull3 makeEmpty]()</h3>
+		<h3>[method:QuickHull3 makeEmpty]()</h3>
 
-    <div>Makes this convex hull empty.</div>
+		<div>Makes this convex hull empty.</div>
 
-    <h3>[method:QuickHull3 addVertexToFace]( [page:Vertex vertex], [page:Face face]  )</h3>
+		<h3>[method:QuickHull3 addVertexToFace]( [page:Vertex vertex], [page:Face face]	)</h3>
 		[page:Vertex vertex] - The vetex to add.<br /><br />
 		[page:Face face] - The target face.<br /><br />
 
-    <div>Adds a vertex to the 'assigned' list of vertices and assigns it to the given face.</div>
+		<div>Adds a vertex to the 'assigned' list of vertices and assigns it to the given face.</div>
 
-    <h3>[method:QuickHull3 removeVertexFromFace]( [page:Vertex vertex], [page:Face face]  )</h3>
+		<h3>[method:QuickHull3 removeVertexFromFace]( [page:Vertex vertex], [page:Face face]	)</h3>
 		[page:Vertex vertex] - The vetex to remove.<br /><br />
 		[page:Face face] - The target face.<br /><br />
 
-    <div>Removes a vertex from the 'assigned' list of vertices and from the given face. It also makes sure that the link from 'face' to the first vertex it sees in 'assigned' is linked correctly after the removal.</div>
+		<div>Removes a vertex from the 'assigned' list of vertices and from the given face. It also makes sure that the link from 'face' to the first vertex it sees in 'assigned' is linked correctly after the removal.</div>
 
-		<h3>[method:Vertex removeAllVerticesFromFace]( [page:Face face]  )</h3>
+		<h3>[method:Vertex removeAllVerticesFromFace]( [page:Face face]	)</h3>
 		[page:Face face] - The given face.<br /><br />
 
-    <div>Removes all the visible vertices that a given face is able to see which are stored in the 'assigned' vertext list.</div>
+		<div>Removes all the visible vertices that a given face is able to see which are stored in the 'assigned' vertext list.</div>
 
-		<h3>[method:QuickHull3 deleteFaceVertices]( [page:Face face], [page:Face absorbingFace]  )</h3>
+		<h3>[method:QuickHull3 deleteFaceVertices]( [page:Face face], [page:Face absorbingFace]	)</h3>
 		[page:Face face] - The given face.<br /><br />
 		[page:Face absorbingFace] - An optional face that tries to absorb the vertices of the first face.<br /><br />
 
-    <div>Removes all the visible vertices that 'face' is able to see.
+		<div>Removes all the visible vertices that 'face' is able to see.
 			<ul>
 				<li>If 'absorbingFace' doesn't exist, then all the removed vertices will be added to the 'unassigned' vertex list.</li>
 				<li>If 'absorbingFace' exists, then this method will assign all the vertices of 'face' that can see 'absorbingFace'.</li>
@@ -100,26 +100,26 @@
 			</ul>
 		</div>
 
-		<h3>[method:QuickHull3 resolveUnassignedPoints]( [page:Array newFaces]  )</h3>
+		<h3>[method:QuickHull3 resolveUnassignedPoints]( [page:Array newFaces]	)</h3>
 		[page:Face newFaces] - An array of new faces.<br /><br />
 
-    <div>Reassigns as many vertices as possible from the unassigned list to the new faces.</div>
+		<div>Reassigns as many vertices as possible from the unassigned list to the new faces.</div>
 
 		<h3>[method:Object computeExtremes]()</h3>
 
-    <div>Computes the extremes values (min/max vectors) which will be used to compute the inital hull.</div>
+		<div>Computes the extremes values (min/max vectors) which will be used to compute the inital hull.</div>
 
 		<h3>[method:QuickHull3 computeInitialHull]()</h3>
 
-    <div>Computes the initial simplex assigning to its faces all the points that are candidates to form part of the hull.</div>
+		<div>Computes the initial simplex assigning to its faces all the points that are candidates to form part of the hull.</div>
 
 		<h3>[method:QuickHull3 reindexFaces]()</h3>
 
-    <div>Removes inactive (e.g. deleted) faces from the internal face list.</div>
+		<div>Removes inactive (e.g. deleted) faces from the internal face list.</div>
 
 		<h3>[method:Vertex nextVertexToAdd]()</h3>
 
-    <div>Finds the next vertex to create faces with the current hull.
+		<div>Finds the next vertex to create faces with the current hull.
 			<ul>
 				<li>Let the initial face be the first face existing in the 'assigned' vertex list.</li>
 				<li>If a face doesn't exist then return since there're no vertices left.</li>
@@ -127,13 +127,13 @@
 			</ul>
 		</div>
 
-		<h3>[method:QuickHull3 computeHorizon]( [page:Vector3 eyePoint], [page:HalfEdge crossEdge], [page:Face face], [page:Array horizon]  )</h3>
+		<h3>[method:QuickHull3 computeHorizon]( [page:Vector3 eyePoint], [page:HalfEdge crossEdge], [page:Face face], [page:Array horizon]	)</h3>
 		[page:Vector3 eyePoint] - The 3D-coordinates of a point.<br /><br />
 		[page:HalfEdge crossEdge] - The edge used to jump to the current face.<br /><br />
 		[page:Face face] - The current face being tested.<br /><br />
 		[page:Array horizon] - The edges that form part of the horizon in CCW order.<br /><br />
 
-    <div>Computes a chain of half edges in CCW order called the 'horizon'. For an edge to be part of the horizon it must join a face that can see 'eyePoint' and a face that cannot see 'eyePoint'.</div>
+		<div>Computes a chain of half edges in CCW order called the 'horizon'. For an edge to be part of the horizon it must join a face that can see 'eyePoint' and a face that cannot see 'eyePoint'.</div>
 
 		<h3>[method:HalfEdge addAdjoiningFace]( [page:Vertex eyeVertex], [page:HalfEdge horizonEdge] )</h3>
 		[page:Vertex eyeVertex] - The vertex that is added to the hull.<br /><br />
@@ -170,6 +170,6 @@
 
 		<h2>Source</h2>
 
-		[link:https://github.com/mrdoob/three.js/blob/master/src/[path].js src/[path].js]
+		[link:https://github.com/mrdoob/three.js/blob/master/examples/js/QuickHull.js examples/js/QuickHull.js]
 	</body>
 </html>
