+++ docs/examples/quickhull/VertexList.html
@@ -2,7 +2,7 @@
 <html lang="en">
 	<head>
 		<meta charset="utf-8" />
-		<base href="../../../" />
+		<base href="../../" />
 		<script src="list.js"></script>
 		<script src="page.js"></script>
 		<link type="text/css" rel="stylesheet" href="page.css" />
@@ -24,32 +24,32 @@
 
 		<h2>Properties</h2>
 
-    <h3>[property:Vertex head]</h3>
-    <div>
-      Reference to the first vertex of the linked list. Default is null.
-    </div>
+		<h3>[property:Vertex head]</h3>
+		<div>
+			Reference to the first vertex of the linked list. Default is null.
+		</div>
 
 		<h3>[property:Vertex tail]</h3>
 		<div>
 			Reference to the last vertex of the linked list. Default is null.
 		</div>
 
-    <h2>Methods</h2>
+		<h2>Methods</h2>
 
-    <h3>[method:Vertex first]()</h3>
+		<h3>[method:Vertex first]()</h3>
 		<div>Returns the head reference.</div>
 
-    <h3>[method:Vertex last]()</h3>
-    <div>Returns the tail reference.</div>
+		<h3>[method:Vertex last]()</h3>
+		<div>Returns the tail reference.</div>
 
-    <h3>[method:VertexList clear]()</h3>
-    <div>Clears the linked list.</div>
+		<h3>[method:VertexList clear]()</h3>
+		<div>Clears the linked list.</div>
 
-    <h3>[method:VertexList insertBefore]( [page:Vertex target], [page:Vertex vertex] )</h3>
-    [page:Vertex target] - The target vertex. It's assumed that this vertex belongs to the linked list.<br /><br />
-    [page:Vertex vertex] - The vertex to insert.<br /><br />
+		<h3>[method:VertexList insertBefore]( [page:Vertex target], [page:Vertex vertex] )</h3>
+		[page:Vertex target] - The target vertex. It's assumed that this vertex belongs to the linked list.<br /><br />
+		[page:Vertex vertex] - The vertex to insert.<br /><br />
 
-    <div>Inserts a vertex <strong>before</strong> a target vertex.</div>
+		<div>Inserts a vertex <strong>before</strong> a target vertex.</div>
 
 		<h3>[method:VertexList insertAfter]( [page:Vertex target], [page:Vertex vertex] )</h3>
 		[page:Vertex target] - The target vertex. It's assumed that this vertex belongs to the linked list.<br /><br />
@@ -84,6 +84,6 @@
 
 		<h2>Source</h2>
 
-		[link:https://github.com/mrdoob/three.js/blob/master/src/[path].js src/[path].js]
+		[link:https://github.com/mrdoob/three.js/blob/master/examples/js/QuickHull.js examples/js/QuickHull.js]
 	</body>
 </html>
