+++ docs/examples/quickhull/VertexNode.html
@@ -2,7 +2,7 @@
 <html lang="en">
 	<head>
 		<meta charset="utf-8" />
-		<base href="../../../" />
+		<base href="../../" />
 		<script src="list.js"></script>
 		<script src="page.js"></script>
 		<link type="text/css" rel="stylesheet" href="page.css" />
@@ -25,10 +25,10 @@
 
 		<h2>Properties</h2>
 
-    <h3>[property:Vector3 point]</h3>
-    <div>
-      A point (x, y, z) in 3D space. Default is undefined.
-    </div>
+		<h3>[property:Vector3 point]</h3>
+		<div>
+			A point (x, y, z) in 3D space. Default is undefined.
+		</div>
 
 		<h3>[property:Vertex prev]</h3>
 		<div>
@@ -47,6 +47,6 @@
 
 		<h2>Source</h2>
 
-		[link:https://github.com/mrdoob/three.js/blob/master/src/[path].js src/[path].js]
+		[link:https://github.com/mrdoob/three.js/blob/master/examples/js/QuickHull.js examples/js/QuickHull.js]
 	</body>
 </html>
