+++ examples/webgl_geometry_convex.html
@@ -30,6 +30,8 @@
 
 		<script src="../build/three.js"></script>
 		<script src="js/controls/OrbitControls.js"></script>
+		<script src="js/QuickHull.js"></script>
+		<script src="js/geometries/ConvexGeometry.js"></script>
 		<script src="js/Detector.js"></script>
 		<script src="js/libs/stats.min.js"></script>
