+++ src/constants.js
@@ -123,5 +123,3 @@ export var RGBM16Encoding = 3005;
 export var RGBDEncoding = 3006;
 export var BasicDepthPacking = 3200;
 export var RGBADepthPacking = 3201;
-export var Visible = 0;
-export var Deleted = 1;
